+++ src/main.rs
@@ -101,14 +101,38 @@ impl Buildpack for PythonBuildpack {
         } = &requested_python_version
         {
             log_warning(
-                "Support for Python 3.9 is deprecated",
+                "Support for Python 3.9 is ending soon",
                 formatdoc! {"
-                    Python 3.9 will reach its upstream end-of-life in October 2025,
-                    at which point it will no longer receive security updates:
+                    Python 3.9 reached its upstream end-of-life on 31st October 2025,
+                    and so no longer receives security updates:
                     https://devguide.python.org/versions/#supported-versions
 
                     As such, support for Python 3.9 will be removed from this
                     buildpack on 7th January 2026.
+
+                    Upgrade to a newer Python version as soon as possible, by
+                    changing the version in your {origin} file.
+
+                    For more information, see:
+                    https://devcenter.heroku.com/articles/python-support#supported-python-versions
+                "},
+            );
+        } else if let RequestedPythonVersion {
+            major: 3,
+            minor: 10,
+            origin,
+            ..
+        } = &requested_python_version
+        {
+            log_warning(
+                "Support for Python 3.10 is deprecated",
+                formatdoc! {"
+                    Python 3.10 will reach its upstream end-of-life in October 2026,
+                    at which point it will no longer receive security updates:
+                    https://devguide.python.org/versions/#supported-versions
+
+                    As such, support for Python 3.10 will be removed from this
+                    buildpack on 6th January 2027.
 
                     Upgrade to a newer Python version as soon as possible, by
                     changing the version in your {origin} file.