+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Deprecated support for Python 3.10. ([#467](https://github.com/heroku/buildpacks-python/pull/467))
 - Unpinned Poetry's `dulwich` version, since the upstream incompatibility with older Python has been fixed. ([#464](https://github.com/heroku/buildpacks-python/pull/464))
 
 ## [2.7.4] - 2025-11-14